Refactor norm_range to context manager instead of permanent set
