+++ zfit/util/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import print_function, division, absolute_import
+
+__all__ = ['AppendDocstring']
+
+from tensorflow.python.ops.distributions.util import AppendDocstring
