+++ zfit/util/container.py
