+++ zfit/util/diverse.py
