+++ zfit/util/exception.py
