--- zfit/utils/__init__.py
@@ -1,1 +1,0 @@
-from __future__ import print_function, division, absolute_import
